Integrate filter, interceptor and AOP.
Difference between filter and interceptor:
1. Filters has a wider range, because they are included in a Servlet container like tomcat, they apply to all HTTP requests, including static resources (CSS, JS, images), APIs, and JSP files.
2. Interceptor is inside a web application(Springboot) normally.
Since an interface will get into a container first then the web application, so the cost of time of that interface will be greater in a filter than an interceptor.
We can just choose one from the three methods to handle cross-cutting concerns (like logging, authentication, and performance tracking).